Clarify KDA weight test ownership
