+++ models/demos/deepseek_v3_d_p/tests/kda/checkpoint/test_layer_checkpoint.py
@@ -1,36 +1,20 @@
 # SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
 # SPDX-License-Identifier: Apache-2.0
-"""CPU tests for indexed KDA checkpoint loading."""
+"""CPU tests for loading one KDA layer from an indexed safetensor checkpoint."""
 
 import json
 from pathlib import Path
 
-import torch
 from safetensors.torch import save_file
 
 from models.demos.deepseek_v3_d_p.reference.kda.config import KDAConfig
-from models.demos.deepseek_v3_d_p.reference.kda.weights import required_kda_weight_names, validate_kda_weights
+from models.demos.deepseek_v3_d_p.reference.kda.weights import required_kda_weight_names
 from models.demos.deepseek_v3_d_p.tests.kda.checkpoint_utils import (
     kda_layer_prefix,
     load_kda_layer_state_dict,
     resolve_kda_layer_shards,
 )
-from models.demos.deepseek_v3_d_p.tests.kda.utils import random_weights
-from models.demos.deepseek_v3_d_p.tt.kda.weight_schema import normalize_kda_state_dict
-from tests.ttnn.unit_tests.operations.experimental.kda.kda_test_utils import assert_equal
-
-
-def _full_rank_config(*, num_heads: int = 2) -> KDAConfig:
-    return KDAConfig(
-        hidden_size=64,
-        num_heads=num_heads,
-        head_k_dim=32,
-        head_v_dim=32,
-        conv_kernel_size=4,
-        norm_eps=1e-5,
-        use_full_rank_gate=True,
-        gate_lower_bound=-5.0,
-    )
+from models.demos.deepseek_v3_d_p.tests.kda.utils import make_config, random_weights
 
 
 def _write_indexed_layer(checkpoint_dir: Path, layer_idx: int, config: KDAConfig) -> Path:
@@ -44,7 +28,7 @@ def _write_indexed_layer(checkpoint_dir: Path, layer_idx: int, config: KDAConfig
 
 
 def test_loads_one_indexed_full_rank_kda_layer(tmp_path: Path) -> None:
-    config = _full_rank_config()
+    config = make_config(use_full_rank_gate=True)
     shard = _write_indexed_layer(tmp_path, layer_idx=1, config=config)
 
     assert resolve_kda_layer_shards(tmp_path, 1, config) == (shard,)
@@ -57,7 +41,7 @@ def test_loads_one_indexed_full_rank_kda_layer(tmp_path: Path) -> None:
 
 
 def test_rejects_incomplete_checkpoint_shard_set(tmp_path: Path, expect_error) -> None:
-    config = _full_rank_config()
+    config = make_config(use_full_rank_gate=True)
     _write_indexed_layer(tmp_path, layer_idx=1, config=config).unlink()
 
     with expect_error(FileNotFoundError, "missing complete KDA checkpoint shard"):
@@ -65,7 +49,7 @@ def test_rejects_incomplete_checkpoint_shard_set(tmp_path: Path, expect_error) -
 
 
 def test_rejects_index_missing_required_kda_weight(tmp_path: Path, expect_error) -> None:
-    config = _full_rank_config()
+    config = make_config(use_full_rank_gate=True)
     _write_indexed_layer(tmp_path, layer_idx=1, config=config)
     index_path = tmp_path / "model.safetensors.index.json"
     index = json.loads(index_path.read_text(encoding="utf-8"))
@@ -74,33 +58,3 @@ def test_rejects_index_missing_required_kda_weight(tmp_path: Path, expect_error)
 
     with expect_error(ValueError, "g_proj.weight"):
         resolve_kda_layer_shards(tmp_path, 1, config)
-
-
-def test_weight_validation_reports_exact_name_and_shape(expect_error) -> None:
-    config = _full_rank_config()
-    weights = random_weights(config)
-    weights["q_proj.weight"] = torch.empty(config.q_dim, config.hidden_size + 1)
-
-    with expect_error(ValueError, r"q_proj\.weight shape .* !="):
-        validate_kda_weights(weights, config)
-
-
-def test_normalize_state_dict_trims_kimi_k3_padded_a_log() -> None:
-    config = _full_rank_config(num_heads=96)
-    state_dict = random_weights(config)
-    padded = torch.arange(128, dtype=torch.float32)
-    state_dict["A_log"] = padded
-
-    normalized = normalize_kda_state_dict(state_dict, config)
-
-    assert normalized["A_log"].shape == (1, 1, config.num_heads, 1)
-    assert_equal(padded[: config.num_heads], normalized["A_log"].reshape(-1), name="trimmed A_log")
-
-
-def test_normalize_state_dict_rejects_unsupported_a_log_padding(expect_error) -> None:
-    config = _full_rank_config(num_heads=96)
-    state_dict = random_weights(config)
-    state_dict["A_log"] = torch.arange(127, dtype=torch.float32)
-
-    with expect_error(ValueError, "A_log has 127 entries"):
-        normalize_kda_state_dict(state_dict, config)
